Moved test to proper space
